Add test.sh which tests examples
Very basic test script which grabs all occurances of python at the start
of a line in the README.md. We also add one example which doesn't
conform to this pattern. We also don't pick up the Llama examples yet
because these require credentials, we need to come back to this.